moving all code to plugin.go and tests to plugin_test.go
