+++ docker/Dockerfile.linux.arm
@@ -1,7 +1,7 @@
-FROM alpine:3.6 as alpine
+FROM alpine:3.20 as alpine
 RUN apk add -U --no-cache ca-certificates
 
-FROM alpine:3.6
+FROM alpine:3.20
 ENV GODEBUG netdns=go
 
 COPY --from=alpine /etc/ssl/certs/ca-certificates.crt /etc/ssl/certs/
